refacto: workspace settings
